+++ .gitignore
@@ -29,3 +29,4 @@ yarn-error.log*
 
 # misc
 .DS_Store
+.wakatime-project
