+++ src/components/elements.tsx
@@ -1,25 +1,25 @@
 // Post forms
-import {PostListPage} from './elements/postListPage';
-import {FetchPost, PostFetchStatus} from './elements/postFetch';
-import {PostInfo} from './elements/postInfo';
-import {PostModificationNotes} from './elements/postModNotes';
+import {PostListPage} from './elements/posts/listPage';
+import {FetchPost, PostFetchStatus} from './elements/posts/fetch';
+import {PostInfo} from './elements/posts/info';
+import {PostModificationNotes} from './elements/posts/modNotes';
 
-import {QuestPostForm} from './elements/questPostForm';
-import {QuestPostList} from './elements/questPostList';
-import {QuestPositionOutput} from './elements/questPositionOutput';
-import {QuestPostFetchStatus} from './elements/questFetch';
-import {QuestPostOutput} from './elements/questPostOutput';
+import {QuestPostForm} from './elements/posts/quest/form';
+import {QuestPostList} from './elements/posts/quest/list';
+import {QuestPositionOutput} from './elements/posts/quest/positionOutput';
+import {QuestPostFetchStatus} from './elements/posts/quest/fetch';
+import {QuestPostOutput} from './elements/posts/quest/output';
 
-import {AnalysisPostList} from './elements/analysisPostList';
-import {AnalysisPostOutput} from './elements/analysisPostOutput';
+import {AnalysisPostList} from './elements/posts/analysis/list';
+import {AnalysisPostOutput} from './elements/posts/analysis/output';
 import {
   AnalysisPostFetchStatus,
   CharacterAnalysisPostFetchStatus,
   DragonAnalysisPostFetchStatus,
-} from './elements/analysisFetch';
-import {AnalysisPostFormChara, AnalysisPostFormDragon} from './elements/analysisPostForm';
-import {AnalysisSkillOutput} from './elements/analysisSkill';
-import {PostManageBar, PostManageBarProps} from './elements/postManageBar';
+} from './elements/posts/analysis/fetch';
+import {AnalysisPostFormChara, AnalysisPostFormDragon} from './elements/posts/analysis/form';
+import {AnalysisSkillOutput} from './elements/posts/analysis/skill';
+import {PostManageBar, PostManageBarProps} from './elements/posts/manageBar';
 
 // 3rd party "plugins"
 import {getGoogleUid} from './elements/googleSignin';
@@ -27,11 +27,11 @@ import {getGoogleUid} from './elements/googleSignin';
 // Components
 import {Navigation} from './elements/navigation';
 import {Footer} from './elements/footer';
-import {PageAnchor, scrollToAnchor} from './elements/pageAnchor';
-import {Paginator} from './elements/paginator';
+import {PageAnchor, scrollToAnchor} from './elements/posts/pageAnchor';
+import {Paginator} from './elements/posts/paginator';
 import {ExpressModal} from './elements/modalExpress';
-import {Markdown} from './elements/markdown';
-import {InfoCard} from './elements/infoCard';
+import {Markdown} from './elements/markdown/main';
+import {InfoCard} from './elements/posts/infoCard';
 
 
 export type {
