+++ src/components/elements/posts/quest/fetch.ts
@@ -1,0 +1,7 @@
+import {PostFetchStatus} from '../fetch';
+import {QuestPostGetSuccessResponse} from '../../../../constants/api';
+
+
+export interface QuestPostFetchStatus extends PostFetchStatus {
+  post: QuestPostGetSuccessResponse | null
+}
