--- src/components/elements/questFetch.ts
@@ -1,7 +1,0 @@
-import {PostFetchStatus} from './postFetch';
-import {QuestPostGetSuccessResponse} from '../../constants/api';
-
-
-export interface QuestPostFetchStatus extends PostFetchStatus {
-  post: QuestPostGetSuccessResponse | null
-}
