+++ src/components/pages/constructing.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {useTranslation} from 'react-i18next';
 
 import {PageProps} from './base';
-import {Markdown} from '../elements/markdown';
+import {Markdown} from '../elements/markdown/main';
 
 export const Constructing = ({fnSetTitle}: PageProps) => {
   const {t} = useTranslation();
