+++ src/components/pages/posts/analysis/edit.tsx
@@ -2,7 +2,7 @@ import React, {Dispatch, SetStateAction} from 'react';
 import {useParams} from 'react-router-dom';
 import {useTranslation} from 'react-i18next';
 
-import {PageProps} from './base';
+import {PageProps} from '../../base';
 import {
   AnalysisPostFetchStatus,
   AnalysisPostFormChara,
@@ -10,8 +10,8 @@ import {
   FetchPost,
   getGoogleUid,
   PostFetchStatus,
-} from '../elements';
-import {AnalysisPostType, ApiRequestSender, CharacterAnalysisPost, DragonAnalysisPost} from '../../constants/api';
+} from '../../../elements';
+import {AnalysisPostType, ApiRequestSender, CharacterAnalysisPost, DragonAnalysisPost} from '../../../../constants/api';
 
 
 export const AnalysisEdit = ({fnSetTitle}: PageProps) => {
