+++ src/components/pages/posts/analysis/list.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import {useTranslation} from 'react-i18next';
 
-import {PageProps} from './base';
-import {AnalysisPostList, PostListPage} from '../elements';
-import Path from '../../constants/path';
-import {AnalysisPostListEntry, ApiRequestSender} from '../../constants/api';
+import {PageProps} from '../../base';
+import {AnalysisPostList, PostListPage} from '../../../elements';
+import Path from '../../../../constants/path';
+import {AnalysisPostListEntry, ApiRequestSender} from '../../../../constants/api';
 
 
 export const AnalysisList = ({fnSetTitle}: PageProps) => {
