+++ src/components/pages/posts/analysis/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import {PageProps} from './base';
-import {AnalysisPostOutput} from '../elements';
+import {PageProps} from '../../base';
+import {AnalysisPostOutput} from '../../../elements';
 
 export const AnalysisPage = ({fnSetTitle}: PageProps) => {
   return <AnalysisPostOutput fnSetTitle={fnSetTitle}/>;
