+++ src/components/pages/posts/quest/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import {PageProps} from './base';
-import {QuestPostOutput} from '../elements';
+import {PageProps} from '../../base';
+import {QuestPostOutput} from '../../../elements';
 
 export const QuestPage = ({fnSetTitle}: PageProps) => {
   return <QuestPostOutput fnSetTitle={fnSetTitle}/>;
